+++ core/shared/src/main/scala/cwinter/codecraft/core/api/DroneSpec.scala
@@ -65,13 +65,13 @@ case class DroneSpec(
   def weight = sides + moduleCount
 
   /** Returns the speed of a drone with this spec, measured in units distance per timestep. */
-  def maxSpeed: Double = 30 * (1 + engines) / weight
+  def maxSpeed: Float = 30 * (1 + engines) / weight
 
 
   /** Returns the `radius` for a drone with this spec.
     * The `radius` is used to compute collisions with projectiles or other drones.
     */
-  val radius: Double = {
+  val radius: Float = {
     val radiusBody = 0.5f * 40 / math.sin(math.Pi / sides).toFloat
     radiusBody + Geometry.circumradius(4, sides)
   }
